turboflow: move Fireblocks MPC custody to treasury adapter
Per @RohanNero review, DefiLlama no longer counts EOA/safe-style
balances toward TVL. This PR:

- Reverts `projects/turboflow/index.js` to bridge-contract-only TVL
  (the original #16949 scope), removing the Fireblocks vault owners
  that had been added in earlier commits on this branch. Methodology
  text updated to point to the treasury adapter.
- Adds `projects/treasury/turboflow.js`, a new treasury adapter that
  reports USDT/USDC balances held by the four protocol-operated
  Fireblocks MPC custody wallets (two BSC + two Solana). Solana reads
  use `computeTokenAccount: true` so balances are batched through
  `getMultipleAccounts` instead of the throttled
  `getTokenAccountsByOwner` path.

Combined the two adapters report the same set of protocol-controlled
USDT/USDC we had been measuring; the split simply routes the EOA-side
holdings into the protocol's treasury entry as requested.
